Minor layout fixes
