+++ src/components/Stats/index.tsx
@@ -1,102 +1,33 @@
 import {
-  Box,
   Card,
   CardBody,
   CardHeader,
   CardProps,
   Flex,
-  Grid,
-  Heading,
   HStack,
   Icon,
-  Stack,
+  Link,
+  SkeletonText,
   Text,
   VStack,
 } from '@chakra-ui/react'
-import { PropsWithChildren, useEffect, useState } from 'react'
-import { useTranslation } from 'react-i18next'
+import { PropsWithChildren } from 'react'
+import { Trans, useTranslation } from 'react-i18next'
 import { IconType } from 'react-icons'
-import { generatePath, Link } from 'react-router-dom'
-import { RefreshIntervalBlocks, RoutePath } from '~constants'
-import { useChainInfo } from '~queries/stats'
-import { Icons } from '~src/theme/components/Icons'
 import { VscGraphLine } from 'react-icons/vsc'
 import { LatestBlocks } from '~components/Stats/LatestBlocks'
 import { RawModal } from '~components/Layout/ShowRawButton'
 import { ChainCosts } from '~components/Stats/ChainDetails/ChainCosts'
 import { ChainInfo } from '~components/Stats/ChainDetails/ChainInfo'
-
-interface IStatsCardProps {
-  title: string
-  description: string
-  link?: string
-  icon: IconType
-}
-
-const StatsCard = ({ title, description, link, icon }: IStatsCardProps) => (
-  <Link to={generatePath(link || '')}>
-    <Card direction={'row'} align={'center'}>
-      <Box px={3}>
-        <Icon as={icon} boxSize={7} />
-      </Box>
-      <Stack spacing={0} py={3}>
-        <CardBody p={0}>
-          <Text fontSize='sm'>{title}</Text>
-          <Text fontWeight={'bold'} fontSize='md'>
-            {description}
-          </Text>
-        </CardBody>
-      </Stack>
-    </Card>
-  </Link>
-)
-
-interface IncrementalStatProps {
-  value: number
-  label: string
-}
-
-const IncrementalStat = ({ value, label }: IncrementalStatProps) => {
-  const [displayNumber, setDisplayNumber] = useState(0)
-  const duration = 2000
-
-  useEffect(() => {
-    let intervalId: NodeJS.Timeout
-    let start = Date.now()
-    const stepTime = duration / value
-
-    intervalId = setInterval(() => {
-      const elapsed = Date.now() - start
-      const remainingTime = duration - elapsed
-      const randomOffset = Math.floor(Math.random() * 10) // Random offset between 0 and 9
-      const nextNumber = Math.min(value, Math.floor(elapsed / stepTime) + randomOffset)
-
-      // Ensure the final number is the target number
-      if (remainingTime <= stepTime) {
-        setDisplayNumber(value)
-        clearInterval(intervalId)
-      } else {
-        setDisplayNumber(nextNumber)
-      }
-    }, stepTime)
-
-    return () => clearInterval(intervalId)
-  }, [value, duration])
-
-  return (
-    <Flex direction={'column'} align={'center'}>
-      <Box>
-        <Heading>{displayNumber}</Heading>
-      </Box>
-      <Text fontSize={'2xl'}>{label}</Text>
-    </Flex>
-  )
-}
+import { StatsCards } from '~components/Stats/ChainDetails/StatsCards'
+import { TxCosts } from '~components/Stats/ChainDetails/TxCosts'
 
 interface StatisticsCardProps {
   title: string
   icon: IconType
   raw?: object
+  readMore?: string
+  isLoading?: boolean
 }
 
 export const StatisticsCardWrapper = ({
@@ -104,6 +35,8 @@ export const StatisticsCardWrapper = ({
   icon,
   raw,
   children,
+  readMore,
+  isLoading,
   ...rest
 }: StatisticsCardProps & PropsWithChildren & CardProps) => (
   <Card flex='1' w={'full'} {...rest}>
@@ -114,93 +47,40 @@ export const StatisticsCardWrapper = ({
       </HStack>
       {!!raw && <RawModal color={'lightText'} fontSize={'xs'} obj={raw} />}
     </CardHeader>
-    <CardBody>{children}</CardBody>
+    <CardBody>
+      {isLoading && <SkeletonText noOfLines={3} spacing={3} skeletonHeight={3} py={2} />}
+      {!isLoading && (
+        <Flex h={'full'} direction={'column'} justify={'space-between'}>
+          {children}
+          {readMore && (
+            <Flex direction={'column'} align={'end'}>
+              <Text as={Link} fontSize={'sm'} href={readMore}>
+                <Trans i18nKey={'read_more'}>Read more</Trans>
+              </Text>
+            </Flex>
+          )}
+        </Flex>
+      )}
+    </CardBody>
   </Card>
 )
 
 const Stats = () => {
-  const { data: stats } = useChainInfo({
-    refetchInterval: RefreshIntervalBlocks,
-  })
   const { t } = useTranslation()
-
-  if (!stats) return null
-
-  const averageBlockTime = Number((stats?.blockTime[0] || 0) / 1000).toFixed(1)
-
-  const statsCards: IStatsCardProps[] = [
-    {
-      title: t('stats.average_block_time'),
-      description: t('stats.seconds', {
-        count: parseFloat(averageBlockTime),
-      }),
-      link: RoutePath.BlocksList,
-      icon: Icons.ClockIcon,
-    },
-    {
-      title: t('stats.block_height'),
-      description: t('stats.blocks', {
-        count: stats.height,
-        defaultValue: '{{count}} blocks',
-      }),
-      link: RoutePath.BlocksList,
-      icon: Icons.BlockIcon,
-    },
-    {
-      title: t('stats.transactions_count', { defaultValue: 'Transactions count' }),
-      description: t('stats.transactions', {
-        count: stats.transactionCount,
-        defaultValue: '{{count}} transactions',
-      }),
-      link: RoutePath.TransactionsList,
-      icon: Icons.TxIcon,
-    },
-    {
-      title: t('stats.validators_count', { defaultValue: 'Validators count' }),
-      description: t('stats.validators', {
-        count: stats.validatorCount,
-        defaultValue: '{{count}} validators',
-      }),
-      link: RoutePath.Validators,
-      icon: Icons.ValidatorIcon,
-    },
-  ]
-
-  const circularStats: IncrementalStatProps[] = [
-    {
-      label: t('stats.organizations', { defaultValue: 'Organizations' }),
-      value: stats.organizationCount,
-    },
-    {
-      label: t('stats.elections', { defaultValue: 'Elections' }),
-      value: stats.electionCount,
-    },
-    {
-      label: t('stats.votes', { defaultValue: 'Votes' }),
-      value: stats.voteCount,
-    },
-  ]
 
   return (
     <Flex direction={'column'} gap={4}>
-      <Grid templateColumns={{ base: 'repeat(1, 1fr)', sm: 'repeat(2, 1fr)', lg: 'repeat(4, 1fr)' }} gap={4} mb={8}>
-        {statsCards.map((card, i) => (
-          <StatsCard key={i} {...card} />
-        ))}
-      </Grid>
-
-      <Flex w={'full'} justify={'space-around'} wrap={'wrap'}>
-        {circularStats.map((card, i) => (
-          <IncrementalStat key={i} {...card} />
-        ))}
-      </Flex>
+      <StatsCards />
       <Flex direction={{ base: 'column-reverse', lg: 'row' }} alignItems='start' gap={4}>
-        <StatisticsCardWrapper minH={'530px'} title={t('stats.latest_blocks')} icon={VscGraphLine}>
+        <StatisticsCardWrapper flex='1' minH={'530px'} title={t('stats.latest_blocks')} icon={VscGraphLine}>
           <LatestBlocks />
         </StatisticsCardWrapper>
-        <VStack>
+        <VStack w={'full'} flex='2'>
           <ChainInfo />
-          <ChainCosts />
+          <Flex w={'full'}>
+            <ChainCosts />
+            <TxCosts />
+          </Flex>
         </VStack>
       </Flex>
     </Flex>
